requirements.txt
kaggle
pandas
python-dotenv
numpy